+++ packages/web-app-files/src/components/AppBar/AppBar.vue
@@ -117,7 +117,7 @@
 <script>
 import { mapActions, mapGetters, mapState, mapMutations } from 'vuex'
 import pathUtil from 'path'
-import { useRouter } from '../../composables'
+import { useRouter } from 'web-pkg/src/composables'
 
 import Mixins from '../../mixins'
 import MixinFileActions, { EDITOR_MODE_CREATE } from '../../mixins/fileActions'
