+++ packages/web-app-files/src/composables/index.ts
@@ -1,5 +1,3 @@
-export * from './router'
-export * from './store'
 export * from './fileListHeaderPosition'
 export * from './pagination'
 export * from './sort'
