+++ packages/web-pkg/src/composables/localStorage/index.ts
@@ -1,0 +1,1 @@
+export * from './useLocalStorage'
