+++ packages/web-pkg/tests/unit/composables/store/useMutationSubscription.spec.ts
@@ -1,5 +1,5 @@
 import { MutationPayload } from 'vuex'
-import { useMutationSubscription, useStore } from '../../../../src/composables'
+import { useMutationSubscription, useStore } from 'web-pkg/src/composables'
 import { createWrapper } from './spec'
 
 describe('subscribeMutation', () => {
